Add median performance test, update `TIMES.md`
Add helper `getRandomNumberArrays`, replace similar performance synthesis with it
