+++ TIMES.md
@@ -1,36 +1,41 @@
 ## Nanoutils method's time
 | Method  | Lib   |    10000 |  100000 |  1000000 |
 | ------- | ----- | -------- | ------- | -------- |
-| allPass | nano  |   0.04ms |  0.12ms |   1.26ms |
-|         | ramda |   0.39ms |  4.06ms |  45.76ms |
-|         | diff  |  -0.35ms | -3.94ms | -44.50ms |
-|         |       |  -90.12% | -97.08% |  -97.24% |
+| allPass | nano  |   0.04ms |  0.11ms |   1.18ms |
+|         | ramda |   0.36ms |  3.74ms |  42.17ms |
+|         | diff  |  -0.32ms | -3.63ms | -40.99ms |
+|         |       |  -89.66% | -96.99% |  -97.19% |
 | ------- | ----- | -------- | ------- | -------- |
-| concat  | nano  |   0.00ms |  0.02ms |   0.82ms |
-|         | ramda |   0.00ms |  0.02ms |   0.75ms |
-|         | diff  |  +0.00ms | +0.00ms |  +0.07ms |
-|         |       |  +55.32% |  +5.19% |   +8.85% |
+| concat  | nano  |   0.00ms |  0.02ms |   0.77ms |
+|         | ramda |   0.00ms |  0.02ms |   0.69ms |
+|         | diff  |  +0.00ms | +0.00ms |  +0.08ms |
+|         |       |  +46.45% |  +9.22% |  +11.75% |
 | ------- | ----- | -------- | ------- | -------- |
-| mean    | nano  |   0.04ms |  0.25ms |   2.04ms |
-|         | ramda |   0.09ms |  0.69ms |   6.47ms |
-|         | diff  |  -0.05ms | -0.44ms |  -4.43ms |
-|         |       |  -54.56% | -64.10% |  -68.49% |
+| mean    | nano  |   0.04ms |  0.22ms |   1.92ms |
+|         | ramda |   0.08ms |  0.63ms |   6.22ms |
+|         | diff  |  -0.04ms | -0.40ms |  -4.30ms |
+|         |       |  -54.20% | -64.34% |  -69.12% |
 | ------- | ----- | -------- | ------- | -------- |
-| sum     | nano  |   0.02ms |  0.19ms |   1.94ms |
-|         | ramda |   0.06ms |  0.63ms |   5.97ms |
-|         | diff  |  -0.04ms | -0.44ms |  -4.03ms |
-|         |       |  -67.18% | -69.14% |  -67.51% |
+| median  | nano  |   0.33ms |  4.62ms |  49.47ms |
+|         | ramda |   0.30ms |  4.35ms |  52.01ms |
+|         | diff  |  +0.04ms | +0.27ms |  -2.53ms |
+|         |       |  +12.27% |  +6.13% |   -4.87% |
 | ------- | ----- | -------- | ------- | -------- |
-| zipWith | nano  |   0.05ms |  0.09ms |   1.95ms |
-|         | ramda |   0.01ms |  0.07ms |   1.87ms |
-|         | diff  |  +0.04ms | +0.01ms |  +0.07ms |
-|         |       | +479.37% | +19.02% |   +3.88% |
+| sum     | nano  |   0.02ms |  0.19ms |   1.85ms |
+|         | ramda |   0.06ms |  0.61ms |   5.67ms |
+|         | diff  |  -0.04ms | -0.42ms |  -3.82ms |
+|         |       |  -66.21% | -69.45% |  -67.32% |
+| ------- | ----- | -------- | ------- | -------- |
+| zipWith | nano  |   0.05ms |  0.08ms |   1.86ms |
+|         | ramda |   0.01ms |  0.07ms |   1.82ms |
+|         | diff  |  +0.04ms | +0.01ms |  +0.05ms |
+|         |       | +489.24% | +12.79% |   +2.54% |
 | ------- | ----- | -------- | ------- | -------- |
 
 | Method | Lib   |        0% |       50% |      100% |
 | ------ | ----- | --------- | --------- | --------- |
-| union  | nano  |   38.09ms |   24.70ms |   17.54ms |
-|        | ramda |    1.04ms |    0.91ms |    0.68ms |
-|        | diff  |  +37.05ms |  +23.80ms |  +16.86ms |
-|        |       | +3546.38% | +2623.39% | +2481.72% |
+| union  | nano  |   35.48ms |   22.49ms |   15.42ms |
+|        | ramda |    1.00ms |    0.82ms |    0.64ms |
+|        | diff  |  +34.48ms |  +21.67ms |  +14.78ms |
+|        |       | +3439.09% | +2646.30% | +2292.84% |
 | ------ | ----- | --------- | --------- | --------- |
