+++ lib/mean/mean.performance.js
@@ -1,1 +1,3 @@
-module.exports = require('../sum/sum.performance')
+var { getRandomNumberArrays } = require('../_internal/helpers/performance')
+
+module.exports = getRandomNumberArrays
